requirements.txt
pydantic
requests
neo4j
python-dotenv
